Removed the x86 and Arm publish profiles as they are not supported under Microsoft.DotNet.ILCompiler also updated the readme to reflect this change.
Moved all types for the compiler into System and Internal folders to allow for better readability of the main ZeroSharp.cs file.
