+++ efi-no-runtime/Internal/Runtime/CompilerHelpers/StartupCodeHelpers.cs
@@ -1,0 +1,25 @@
+﻿namespace Internal.Runtime.CompilerHelpers
+{
+    using System;
+    using System.Runtime;
+    // A class that the compiler looks for that has helpers to initialize the
+    // process. The compiler can gracefully handle the helpers not being present,
+    // but the class itself being absent is unhandled. Let's add an empty class.
+    class StartupCodeHelpers
+    {
+        // A couple symbols the generated code will need we park them in this class
+        // for no particular reason. These aid in transitioning to/from managed code.
+        // Since we don't have a GC, the transition is a no-op.
+        [RuntimeExport("RhpReversePInvoke")]
+        static void RhpReversePInvoke(IntPtr frame) { }
+        [RuntimeExport("RhpReversePInvokeReturn")]
+        static void RhpReversePInvokeReturn(IntPtr frame) { }
+        [RuntimeExport("RhpPInvoke")]
+        static void RhpPInvoke(IntPtr frame) { }
+        [RuntimeExport("RhpPInvokeReturn")]
+        static void RhpPInvokeReturn(IntPtr frame) { }
+
+        [RuntimeExport("RhpFallbackFailFast")]
+        static void RhpFallbackFailFast() { while (true) ; }
+    }
+}
