+++ efi-no-runtime/Properties/PublishProfiles/ARM64Release.pubxml
@@ -20,5 +20,6 @@ https://go.microsoft.com/fwlink/?LinkID=208121.
 	  <NoStdLib>true</NoStdLib>
 	  <NoConfig>true</NoConfig>
 	  <BuildVHDX>true</BuildVHDX>
+	  <PlatformTarget>arm64</PlatformTarget>
   </PropertyGroup>
 </Project>
