--- efi-no-runtime/Properties/PublishProfiles/ARMDebug.pubxml
@@ -1,24 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<!--
-https://go.microsoft.com/fwlink/?LinkID=208121.
--->
-<Project>
-  <PropertyGroup>
-    <Configuration>Debug</Configuration>
-    <Platform>Any CPU</Platform>
-    <PublishDir>.\Publish\ARM</PublishDir>
-    <PublishProtocol>FileSystem</PublishProtocol>
-    <_TargetId>ArmDebug</_TargetId>
-    <TargetFramework>net7.0</TargetFramework>
-    <RuntimeIdentifier>win-arm</RuntimeIdentifier>
-    <SelfContained>true</SelfContained>
-    <PublishSingleFile>false</PublishSingleFile>
-    <PublishReadyToRun>true</PublishReadyToRun>
-	  <EntryPointSymbol>EfiMain</EntryPointSymbol>
-	  <LinkerSubsystem>EFI_APPLICATION</LinkerSubsystem>
-	  <PublishAot>true</PublishAot>
-	  <NoStdLib>true</NoStdLib>
-	  <NoConfig>true</NoConfig>
-	  <BuildVHDX>true</BuildVHDX>
-  </PropertyGroup>
-</Project>
