--- efi-no-runtime/Properties/PublishProfiles/x86Debug.pubxml
@@ -1,24 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<!--
-https://go.microsoft.com/fwlink/?LinkID=208121.
--->
-<Project>
-  <PropertyGroup>
-    <Configuration>Debug</Configuration>
-    <Platform>Any CPU</Platform>
-    <PublishDir>.\Publish\x86</PublishDir>
-    <PublishProtocol>FileSystem</PublishProtocol>
-    <_TargetId>x86Debug</_TargetId>
-    <TargetFramework>net7.0</TargetFramework>
-    <RuntimeIdentifier>win-x86</RuntimeIdentifier>
-    <SelfContained>true</SelfContained>
-    <PublishSingleFile>false</PublishSingleFile>
-    <PublishReadyToRun>true</PublishReadyToRun>
-	  <EntryPointSymbol>EfiMain</EntryPointSymbol>
-	  <LinkerSubsystem>EFI_APPLICATION</LinkerSubsystem>
-	  <PublishAot>true</PublishAot>
-	  <NoStdLib>true</NoStdLib>
-	  <NoConfig>true</NoConfig>
-	  <BuildVHDX>true</BuildVHDX>
-  </PropertyGroup>
-</Project>
