+++ efi-no-runtime/System/AppContext.cs
@@ -1,0 +1,7 @@
+﻿namespace System
+{
+    public class AppContext
+    {
+        public static void SetData(string s, object o) { }
+    }
+}
