+++ efi-no-runtime/System/Array.cs
@@ -1,0 +1,6 @@
+﻿namespace System
+{
+    public abstract class Array { }
+
+    class Array<T> : Array { }
+}
