+++ efi-no-runtime/System/Attribute.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public class Attribute { }
+}
