+++ efi-no-runtime/System/AttributeTargets.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public enum AttributeTargets { }
+}
