+++ efi-no-runtime/System/AttributeUsageAttribute.cs
@@ -1,0 +1,9 @@
+﻿namespace System
+{
+    public sealed class AttributeUsageAttribute : Attribute
+    {
+        public AttributeUsageAttribute(AttributeTargets validOn) { }
+        public bool AllowMultiple { get; set; }
+        public bool Inherited { get; set; }
+    }
+}
