+++ efi-no-runtime/System/Boolean.cs
@@ -1,0 +1,6 @@
+﻿namespace System
+{
+    // The layout of primitive types is special cased because it would be recursive.
+    // These really don't need any fields to work.
+    public struct Boolean { }
+}
