+++ efi-no-runtime/System/Byte.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct Byte { }
+}
