+++ efi-no-runtime/System/Char.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct Char { }
+}
