+++ efi-no-runtime/System/Delegate.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public abstract class Delegate { }
+}
