+++ efi-no-runtime/System/Double.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct Double { }
+}
