+++ efi-no-runtime/System/Enum.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public abstract class Enum : ValueType { }
+}
