+++ efi-no-runtime/System/Int16.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct Int16 { }
+}
