+++ efi-no-runtime/System/Int32.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct Int32 { }
+}
