+++ efi-no-runtime/System/Int64.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct Int64 { }
+}
