+++ efi-no-runtime/System/IntPtr.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct IntPtr { }
+}
