+++ efi-no-runtime/System/MulticastDelegate.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public abstract class MulticastDelegate : Delegate { }
+}
