+++ efi-no-runtime/System/Nullable.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct Nullable<T> where T : struct { }
+}
