+++ efi-no-runtime/System/Object.cs
@@ -1,0 +1,10 @@
+﻿namespace System
+{
+    public class Object
+    {
+#pragma warning disable 169
+        // The layout of object is a contract with the compiler.
+        private IntPtr m_pMethodTable;
+#pragma warning restore 169
+    }
+}
