+++ efi-no-runtime/System/Runtime/CompilerServices/RuntimeFeature.cs
@@ -1,0 +1,8 @@
+﻿namespace System.Runtime.CompilerServices
+{
+    public static class RuntimeFeature
+    {
+        public const string UnmanagedSignatureCallingConvention = nameof(UnmanagedSignatureCallingConvention);
+    }
+
+}
