+++ efi-no-runtime/System/Runtime/CompilerServices/RuntimeHelpers.cs
@@ -1,0 +1,7 @@
+﻿namespace System.Runtime.CompilerServices
+{
+    public class RuntimeHelpers
+    {
+        public static unsafe int OffsetToStringData => sizeof(IntPtr) + sizeof(int);
+    }
+}
