+++ efi-no-runtime/System/Runtime/InteropServices/CharSet.cs
@@ -1,0 +1,10 @@
+﻿namespace System.Runtime.InteropServices
+{
+    internal enum CharSet
+    {
+        None = 1,       // User didn't specify how to marshal strings.
+        Ansi = 2,       // Strings should be marshalled as ANSI 1 byte chars.
+        Unicode = 3,    // Strings should be marshalled as Unicode 2 byte chars.
+        Auto = 4,       // Marshal Strings in the right way for the target system.
+    }
+}
