+++ efi-no-runtime/System/Runtime/InteropServices/LayoutKind.cs
@@ -1,0 +1,9 @@
+﻿namespace System.Runtime.InteropServices
+{
+    internal enum LayoutKind
+    {
+        Sequential = 0, // 0x00000008,
+        Explicit = 2, // 0x00000010,
+        Auto = 3, // 0x00000000,
+    }
+}
