+++ efi-no-runtime/System/Runtime/InteropServices/StructLayoutAttribute.cs
@@ -1,0 +1,9 @@
+﻿namespace System.Runtime.InteropServices
+{
+    sealed class StructLayoutAttribute : Attribute
+    {
+        public StructLayoutAttribute(LayoutKind layoutKind)
+        {
+        }
+    }
+}
