+++ efi-no-runtime/System/Runtime/InteropServices/UnmanagedType.cs
@@ -1,0 +1,4 @@
+﻿namespace System.Runtime.InteropServices
+{
+    public class UnmanagedType { }
+}
