+++ efi-no-runtime/System/Runtime/RuntimeExportAttribute.cs
@@ -1,0 +1,8 @@
+﻿namespace System.Runtime
+{
+    internal sealed class RuntimeExportAttribute : Attribute
+    {
+        public RuntimeExportAttribute(string entry) { }
+    }
+
+}
