+++ efi-no-runtime/System/RuntimeFieldHandle.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct RuntimeFieldHandle { }
+}
