+++ efi-no-runtime/System/RuntimeMethodHandle.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct RuntimeMethodHandle { }
+}
