+++ efi-no-runtime/System/RuntimeTypeHandle.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct RuntimeTypeHandle { }
+}
