+++ efi-no-runtime/System/SByte.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct SByte { }
+}
