+++ efi-no-runtime/System/Single.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct Single { }
+}
