+++ efi-no-runtime/System/String.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public sealed class String { public readonly int Length; }
+}
