+++ efi-no-runtime/System/UInt16.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct UInt16 { }
+}
