+++ efi-no-runtime/System/UInt32.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct UInt32 { }
+}
