+++ efi-no-runtime/System/UInt64.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct UInt64 { }
+}
