+++ efi-no-runtime/System/UIntPtr.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct UIntPtr { }
+}
