+++ efi-no-runtime/System/ValueType.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public abstract class ValueType { }
+}
