+++ efi-no-runtime/System/Void.cs
@@ -1,0 +1,4 @@
+﻿namespace System
+{
+    public struct Void { }
+}
