+++ efi-no-runtime/efinoruntime.csproj
@@ -4,23 +4,26 @@
 		<OutputType>Exe</OutputType>
 		<TargetFramework>net7.0</TargetFramework>
 		<AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-
+		<FileAlignment>512</FileAlignment>
+		<RuntimeMetadataVersion>v4.0.30319</RuntimeMetadataVersion>
 		<NoStdLib>true</NoStdLib>
 		<NoConfig>true</NoConfig>
-		<RuntimeMetadataVersion>v4.0.30319</RuntimeMetadataVersion>
 		<GenerateAssemblyInfo>false</GenerateAssemblyInfo>
 		<GenerateTargetFrameworkAttribute>false</GenerateTargetFrameworkAttribute>
-		<IlcSystemModule>efinoruntime</IlcSystemModule>
 		<EntryPointSymbol>EfiMain</EntryPointSymbol>
 		<LinkerSubsystem>EFI_APPLICATION</LinkerSubsystem>
+		<LinkerMachineType>EFI_</LinkerMachineType>
 		<PublishAot>true</PublishAot>
-		
-		
+		<AssemblyExt>.EFI</AssemblyExt>
+		<AssemblyName Condition="'$(RuntimeIdentifier)'=='win-arm'">BOOTARM</AssemblyName>
+		<ilcSystemModule Condition="'$(RuntimeIdentifier)'=='win-arm'">BOOTARM</ilcSystemModule>
+		<AssemblyName Condition="'$(RuntimeIdentifier)'=='win-arm64'">BOOTAA64</AssemblyName>
+		<ilcSystemModule Condition="'$(RuntimeIdentifier)'=='win-arm64'">BOOTAA64</ilcSystemModule>
+		<AssemblyName Condition="'$(RuntimeIdentifier)'=='win-x64'">BOOTX64</AssemblyName>
+		<ilcSystemModule Condition="'$(RuntimeIdentifier)'=='win-x64'">BOOTX64</ilcSystemModule>
+		<AssemblyName Condition="'$(RuntimeIdentifier)'=='win-x86'">BOOTIA32</AssemblyName>
+		<ilcSystemModule Condition="'$(RuntimeIdentifier)'=='win-x86'">BOOTIA32</ilcSystemModule>
 	</PropertyGroup>
-	
-	<ItemGroup>
-		<Folder Include="Publish\amd64\" />
-	</ItemGroup>
 	<!--
     EFI follows the Windows ABI so we would need to pretend we're compiling for Windows. Should be possible
     with targetos ILC argument, but I don't have bandwidth to test. Plus we need to run linker somehow after.
@@ -35,9 +38,9 @@
 			<VHDFileName>$(SolutionName).vhdx</VHDFileName>
 			<VHDFilePath>$(VHDFileFolder)$(VHDFileName)</VHDFilePath>
 		</PropertyGroup>
-		<Message Text="Clearing $(VHDFileName)" Importance="high"/>
-		<Exec Command="diskpart /s &quot;$(IntermediateOutputPath)detachvhd.txt&quot;" ContinueOnError="true"/>
-		<Delete Files="$(VHDFilePath)"/>
+		<Message Text="Clearing $(VHDFileName)" Importance="high" />
+		<Exec Command="diskpart /s &quot;$(IntermediateOutputPath)detachvhd.txt&quot;" ContinueOnError="true" />
+		<Delete Files="$(VHDFilePath)" />
 	</Target>
 	
 	<Target Name="BuildVHDX" AfterTargets="CopyNativeBinary" Condition="'$(BuildVHDX)' == 'true'" Inputs="$(PublishDir)$(TargetName).exe" Outputs="$(PublishDir)zerosharp.vhdx">
@@ -45,13 +48,10 @@
 			<VHDFileFolder>$([System.IO.Path]::GetFullPath('$(PublishDir)'))</VHDFileFolder>
 			<VHDFileName>$(SolutionName).vhdx</VHDFileName>
 			<VHDFilePath>$(VHDFileFolder)$(VHDFileName)</VHDFilePath>
-			<BootFile Condition="'$(RuntimeIdentifier)'=='win-arm'">BOOTARM.EFI</BootFile>
-			<BootFile Condition="'$(RuntimeIdentifier)'=='win-arm64'">BOOTAA64.EFI</BootFile>
-			<BootFile Condition="'$(RuntimeIdentifier)'=='win-x64'">BOOTX64.EFI</BootFile>
-			<BootFile Condition="'$(RuntimeIdentifier)'=='win-x86'">BOOTIA32.EFI</BootFile>
 			<DriveLetter>Z</DriveLetter>
-			<SourceFile>$(PublishDir)$(TargetName).exe</SourceFile>
-			<DestinationFiles>$(DriveLetter):\EFI\BOOT\$(BootFile)</DestinationFiles>
+			<!--TargetName-->
+			<SourceFile>$(PublishDir)$(AssemblyName).exe</SourceFile>
+			<DestinationFiles>$(DriveLetter):\EFI\BOOT\$(AssemblyName).efi</DestinationFiles>
 			<AttachVHDLines>
 				create vdisk file=$(VHDFilePath) maximum=500
 				select vdisk file=$(VHDFilePath)
@@ -71,18 +71,19 @@
 		<WriteLinesToFile File="$(IntermediateOutputPath)attachvhd.txt" Overwrite="true" Lines="$(AttachVHDLines)" />
 		<WriteLinesToFile File="$(IntermediateOutputPath)detachvhd.txt" Overwrite="true" Lines="$(DetachVHDLines)" />
 		
-		<CallTarget Targets="CleanVHDX" Condition="$([System.IO.File]::Exists('$(VHDFilePath)'))"/>
+		<CallTarget Targets="CleanVHDX" Condition="$([System.IO.File]::Exists('$(VHDFilePath)'))" />
 		<Message Text="Building VHDX at $(VHDFilePath)." Importance="High" />
 		<Message Text="Creating And Attaching VHD" Importance="High" />
 		<Exec Command="diskpart /s &quot;$(IntermediateOutputPath)attachvhd.txt&quot;" />
-		<Message Text="Created And Attached VHD" Importance="high"/>
-		<Message Text="Copying $([System.IO.Path]::GetFullPath('$(SourceFile)')) to $(DestinationFiles)" Importance="high"/>
+		<Message Text="Created And Attached VHD" Importance="high" />
+		<Message Text="Copying $([System.IO.Path]::GetFullPath('$(SourceFile)')) to $(DestinationFiles)" Importance="high" />
 		<Copy SourceFiles="$([System.IO.Path]::GetFullPath('$(SourceFile)'))" DestinationFiles="$(DestinationFiles)" />
-		<Message Text="Detaching VHD" Importance="high"/>
+		<!--<Copy SourceFiles="$([System.IO.Path]::GetFullPath('$(SourceFile)'))" DestinationFiles="$(DestinationFiles)" />-->
+		<Message Text="Detaching VHD" Importance="high" />
 		<Exec Command="diskpart /s &quot;$(IntermediateOutputPath)detachvhd.txt&quot;" />
-		<Message Text="Detached VHD" Importance="high"/>
-		<Delete Files="$(IntermediateOutputPath)detachvhd.txt"/>
-		<Delete Files="$(IntermediateOutputPath)attachvhd.txt"/>
+		<Message Text="Detached VHD" Importance="high" />
+		<Delete Files="$(IntermediateOutputPath)detachvhd.txt" />
+		<Delete Files="$(IntermediateOutputPath)attachvhd.txt" />
 		<Message Text="Built VHDX at $(VHDFilePath)." Importance="High" />
 	</Target>
